Fix Explore-connect
Fixes #738

The JS that starts the OAuth flow was using a href attribute that's no longer
there now that the connect buttons are <button>s, not <a>s. Now the JS will just
interpolate its own URL based on the procider attribute.

Because this was copied in two places, I made js/locker.js to start
encapsulating shared functionality.

Also removed a <content> tag that isn't valid HTML and whitespaced connect.ejs.