Fix "Connect services" popup in the App Gallery
Closes #745

The App Gallery (and other occurences of an empty modal) was trying to show you
the "Connect more services" popup, but it hadn't been rendered. I moved the
modal rendering into the layout so it'll be available no matter what, and
gave it a more descriptive name on the way.

The JS that actually rendered it seemed overly complicated and is now simpler.
We were explicitly checking that we hadn't already popped it up, but I don't see
any code paths that could possibly make that happen more than once. To be safe,
I left logic like that in but used the cookie we were already checking to make
it more straigtforward.

We were also allowing the modal to render on Explore-connect, which seems weird,
so now it doesn't.

Bonus, fixed whitespace to the 4-space standard we appear headed for.