Choose a random port number for locker and/or mongo if 0 is chosen
This allows us to specify in the config file that locker should choose
random ports.  The test suite is now configured to do this.

This is a bit of a hack, because it doesn't check if the port number is
already in use, so there are possible conflicts and races here.  But it
should be good enough for things to work reliably in CI at our current
pace.