Sample zombie.js integration tests for the locker.
It turns out that the complexity of the locker dashboard is too much for
zombie's rotting undead brain. The level of redirection causes zombie to
lose track of the base of relative URL loads. .less files explode when
the less compiler tries to fish them out of the DOM and turn them into
CSS (this may have something to do with JSDOM's poor support for
styles). And right now it appears to be very difficult to get at the
contents of dynamically-loaded iframes. These tests are very close to
working, but that last issue has to get resolved. Committing to get in
front of other eyeballs.