Start of a smarter collections pipeline.
The pipeline beings with an injector, pushing the data into the pipeline
raw.  It then goes through the first pump that takes a synclet result
and breaks it into an array of changes, this is the complete changeset.
Each entry has a small wrapper of the idr and this is where each pump in
the pipeline (usually a colleciton) can add it's own data.  Each pump
must return the data that is to be passed to the next pump.  This is a
standard waterfall pattern and is very order dependent.