Refactor facebook/lib
This should be a lot of commits, but it came out of a rabbithole.

- Standard linting
- Removed unused functions
  - `getCheckingFeed`
  - `getAlbums`
  - `getTagged`
  - `getObject`
- Renamed nondescriptive functions
  - `getData` -> `getPages`
  - `getOne`  -> `getPage`
- New `getData` DRYs up limit/since logic in several `get*` functions
- Consolidate URL building in `apiUrl`
- Our FQL fetches can use `getFQL` like code outside the library does

Tests started failing because the query parameter order wasn't the same as that
picked by `querystring` (through `url`). JavaScript objects aren't ordered, so
this can't be relied on, but for the moment (and several hundred test runs), it
seems to be consistent. We could do something kludgy like alphebetize them, but
the real fix would be to upgrade `node-fakeweb` to ignore query order.