debug for a weird ijod crasher
The crasher is currently unknown, but likely causes are S3 returning
incorrect data or HTTP parsing returning invalid data.  This adds some
debugging to start to catch it.  It is not a full solution.

Further, this has to bail in the middle of results to not crash, and
finding a clean way to do that is elluding me.