Authentication working, synlets in but untested
Waiting on testing until I can run the synclets, as they are currently not being run.
Had to modify the oauth module, which is not part of the repo, so I need
to figure out the best way to include that. Also, in addition to the oauth flow
Zeo requires an api key that is tied to the domain making the api calls. That is
stored as a constant in lib.js.