Zeo Connector: auth, proxy, zqScore, sleepStats
Modified local oauth module. Syncs zqScore and sleepStats
daily. zqScore is the overall average zqScore, which is updated
daily by Zeo, but here it is synced with an additional date field
to archive. A pid is created using the bedTime field of the earliest
sleep record for the user, which will probably not be duplicated by
another user.