Provide user gating at account creation
When an account goes to actually be created we can now check for gating.
The gating is stored in ijod under the idr of gate:<app id>.  The entry
must have a users field that is an integer of how many users the app can
have.  A value of 0 disables all user creation and a value of -1 is
unlimited.  Any other value is the max user count.  Once the user count
is reached new user creation is disallowed.