Accept photo uploads with no messages
These are starting to look pretty similar and could be refactored. It's also a
little hacky to just send over an empty string (rather than not send that
parameter where there's no `body`), but so far Facebook and Twitter accept it
and deal with it correctly, so it's the easier thing to do.