Search logs back to N hours
After we get one batch of logs, we compare Papertrail's minimum timestamp to our
target and make another call if we need to. Takes a while for times over a
couple hours, but since we have to go based on the IDs they return, I don't
immediately see a way to parallelize it.