parallelize synclet runs for a profile.
taskman now runs all tasks (synclets) for a pid in parallel. It
saves config updates in a list as it goes an applies them all in series
at the end to avoid over-writing each other. That can definitely be
optimized/simplified/improved in the future, but I'd rather get this
out in prod and working first.