Allow shared state across parallel synclets
Now we can prefix special config variables with `_` and they'll get hoisted into
the global config space. During sandboxing, any `_var`s are included in the
synclet namespace, and when the synclet returns a new config, `_var`s are merged
back into the top-level config object.

This allows us to communicate between synclets that somehow depend on each
other. Fitbit, for example, needs to pass the last device sync time to the rest
of the synclets so they know how far to fetch data.

Now, the config looks like:

```
{
  synclet1: {...},
  synclet2: {...},
  _shared1: ...,
  _shared2: ...
}
```

This obviously reintroduces the danger of synclets clobbering each other's data,
so we must do our best to isolate writing of a given variable to one synclet and
leave the rest only to read. If it becomes necessary, we can give each synclet a
writeable section that it made available to other synclets, but ignores their
changes to it. Convention should do for now.