Accept `to` parameter; Refactor profile parsing
On @jparkrr's suggestion, we're going to using `to` and `from` to denote the
profile you're posting an item to and who you're posting it as. Since they'll
have the same format, `parseProfileList` now encapsulates that logic.

The `services` parameter should be considered deprecated.

This happens to clear up some noise we were probably getting in statsd, where we
used to pass things like `singlytesting@tumblr` when we meant just `tumblr`.
I can imaging us wanting to note whether there was a target profile or not to
measure usage, but not in this commit.