cache linkedin profiles app-wide in ijod.
all linkedin profile objects will be cached in ijod so that if A, B and C are
all connected to Q and Q updates her profile, once the first of A, B, or C pull
down that update, the others won't have to make API calls to get it, it will
simply be pulled from the cache.

They can only be cached on a per-app basis for privacy (auth scope) and because
LinkedIn returns different profile ids per-app.

It's set to log both cache hits and misses, so we can see how much we are
saving. Might be good to advertise: Hitting LinkedIn rate limits? Reduce calls
by XX% with Singly!