Use FB Page tokens while posting
If the declared `from` is not the user ID (or empty), we load the user's pages
and scan them for a matching ID. If found, its access token is used. Otherwise,
we let you know you don't have a token for it and what you might be able to do.