Create `og` "type"
This isn't conceptually a type like everything under `/types/*`, but it posts
just the same. We cheat by faking the `:type` parameter from the OG endpoint,
and then use the actual incoming `action` parameter as the `POST` endpoint.

Anything sent to us as `og_*` has the `og_` stripped from it and is passed
directly to Facebook.