Stop using `env.PORT` for apihosts
This was put in to support Heroku (7d4a1e5) but isn't needed anymore. It
interferes with Foreman, which sets `PORT` to 500{0,1,2,...} as it starts
processes. Now we can start a whole hallway suite (apihost, dawg, worker) with
one `foreman start` and everything will be on the ports we've grown to love.

@beaugunderson assures me `PORT` is unused in production :)