Record duration of all request paths
We were only recording the first section of the URL path, for example
`/services/facebook` would become `/services`. Now we'll record the whole path.
In order to not conflate the new timings with the old, the new namespace is
`durations` (was `duration`). When we no longer care about the old data, we can
prune that namespace.

This also installs
[underscore.string](http://epeli.github.com/underscore.string/). It handles
trimming of arbitrary characters (not just whitespace, which is all the built-in
`trim` does) as well as a bundle of other handy string manipulations.