Fix soundcloud/self
This wasn't getting stored because a `map.js` is now required. While I was in
there, I switched `contact` to the preferred `profile`. This will affect
@beaugunderson in production and that's it - I'm actually curious how it
manifests.

For the future, be aware that when a `map.js` is missing, we appear to stick an
`anubis` key into the service discovery endpoint. Haven't tracked down why.

```
{
  "anubis": 0,
  "self": 0
}
```