Add sharing via Twitter Direct Message (DM)
Closes #373

Twitter gets points for making this as easy as it should be. In order for the
access token to get DM access, the developer must include `?flag=dm` during
authorization (will be documented).

I put this at `POST /types/messages`, which will be the endpoint for LinkedIn as
well, and where we can also put a consolidated inbox as we support reading too.