Accept non-GET requests via Shutterfly proxy
Express/Connect appears has removed both the old `req.rawBody` parameter (for
memory reasons) and the ability to add custom parsers for specific MIME types
(for unknown reasons). This reintroduces the `rawBody` field for `POST`s to us
only of type `application/xml` (for minimal impact reasons) so that we can pass
it along to Shutterfly. Since we don't actually care what's in it, I haven't
bothered to parse it.

Beyond that, Shutterfly requires that the app ID be sent along with non-GET
requests. Don't know why, since the access token is presumably already tied to
an app.