Add photos synclet
This is a slightly different pattern from our other photos synclets. Instead of
conflating the albums fetch with the photos fetch, we just reenqueue the photos
task if albums haven't synched yet. If the user actually has 0 albums, this may
cause trouble, but everyone is started with 2 albums upon signup, so we'll deal
with that when we find the guy who deleted them and didn't add anything new. We
could simply set a config flag from the albums synclet saying it's run once and
use that to tell the difference between not-run and empty.

Shutterfly's API doesn't appear to have any concept of paging, so we'll see what
happens when we find a big album.

Their default photo URLs are also cropped, which is sad, and the
[size selection](http://www.shutterfly.com/documentation/api_Proctaserv.sfly) is
obtuse. We can pick a known-sized uncropped image, but the `raw=1` parameter
that is supposed to work has no effect. I'm leaving it alone while I track down
someone from shutterfly to ask about it - their
[support email](mailto:developerhelp@shutterfly.com) bounces. The docs also say
we need to sign the album request, but it's working without. I'll try it to see
if it makes `raw` work.