Sandbox Foursquare synclets
This causes the Foursquare synclets to run in sandboxed/parallel mode, in which
each synclet's config is isolated from the rest. It appears to Just Work™, which
is to say I get the same item counts I did before and there's nothing in the
code that looks like it cares.