Fix broken paging in Meetup synclets
We were trying to keep ourselves on the last incomplete page by subtracting one
from our offset once we discover an empty page, but if there are no items at
all, we ended up looking for page -1. Meetup would throw an error and we'd never
recover.

This simple fix just prevents us from going below 0 again, and cleans up any bad
configs that may exist right now, but we appear to be tracking this page value
in a couple redundant places and mixing it with the idea of a "since" time too.
That could use some further real cleanup.