Fix synclet delay logging
Two problems here. One was dumb - I was checking an array against 0 instead of
the actual time delay. Turns out arrays are never less than 0.

The other is less dumb - as the new task system went out, the shape of the task
objects changed. JavaScript happily went with it, but we started generating
Graphite delay keys that lacked the service name, so all the synclets were being
spammed a level too high.