Rescope synclet delays
I'm moving the new, working timers to a new namespace to avoid the messy data in
the first one. We can clean the other space up whenever we like, but it might be
useful to look at for a sense of how the past week's changes have affected sync
delays, even though synclets with the same names got conflated.