Teach profile verifier to fix the mismatch
With the `--fix` option, the script will save to the KV store any profiles not
found in it yet. It does this simply by calling `allSet` with no updates (it's
really a merge, not a set). The get performed will fall back and load from the
`Profiles` table, and then the save happens on the KV store.