Create profiles on remote pods
If the profile doesn't exist locally and we're not in a `pod` role, we pass the
request over to the pod that's supposed to deal with it (currently hardcoded to
pod `1`). Once that pod has returned successfully, we create our own local entry
with the `pod` column set to indicate the remote location.