Build and test `podClient.getRange`
The client loads the profile in question and either passes it through to `ijod`
directly or makes an HTTP request to the appropriate pod. On the pod side, a
stub `cbEach` is used to accumulate the data to be passed back. Then on the
client side, the original `cbEach` is run over the returned data.

The service side could use some tests, but that's still hard to bootstrap.