Update Anubis tests
Since we're no longer making a local IJOD entry, the `ran` variable was never
getting set by the `dal-fake` match. What we really want to test is that
`nexusClient.batchSmartAdd` completed, which we now do with help from Fakeweb.
Fakeweb doesn't let us run functions, so I added a callback to `anubis.reap`
that anything can use if it's curious about results. In that callback, we can
test that the `nexusClient` returned what we expect, though we're still not
testing any of the actual data logged.