Fixes to Google Proxy to allow binary uploads.
Proxy was parsing as json so allowed the content-type to be passed
through as well as the raw body. In order to support binary data changed
the body to be set to a binary buffer rather than allow request to
default to a utf8 buffer. Changed the middleware code that skips the
bodyparser to be specific for google proxy to also set the encoding for
the rawBody buffer to binary.