Make Foursquare/friends page
This synclet was only ever getting a max of 500 friends; now it'll get all of
them.

It also does a metric ton of extra work getting contact data that has no
business being in contact data or being fetched as part of a core synclet.
That's still happening because we don't want to suddently change our API
content, but it should me moved into other synclets if people need it.

For now, at least the extra work will be done in batches of 100 (which is still
21 requests to Foursquare during the synclet run), but that's better than 500
and should help free the workers.