Make Facebook photos page
I didn't make albums page on the assumption that very few people will have
*that* many of them. We also haven't paged any FQL yet.

I wanted to just update the album's `since` time as we gathered photos, but it
interacts poorly with `limit`. Basically, `limit` is applied first, so on the
second call, we get nothing back.

Instead, I parse the `after` out of Facebook's returned paging URL and use that
on the next call. I wanted to use the whole URL they give back, but it doesn't
include an `access_token` and this way fit better with existing code like
`apiUrl`.