Remove recursion from Facebook page_likes
This will be the only likes synclet that pages forward in time unless we decide
it's important to reverse it. The reason being it's the only source of likes
that lets us page based on timestamp, and it's easier, more readable, and less
likely to have bugs if we just use that to run one direction.